refactor(operator): extract NetworkTab from NetworkSection
Moves NetworkSection's full UI into NetworkTab (§2.11); bootstrap query
moved into NetworkTab so the section is always defaultExpanded=true.
NetworkSection.tsx and its test are deleted. Operator.tsx has
NetworkSection references removed.
